create a helpers.ts file for jest plugin
