+++ packages/knip/src/plugins/jest/index.ts
@@ -1,7 +1,8 @@
 import type { IsPluginEnabled, Plugin, PluginOptions, ResolveConfig, ResolveEntryPaths } from '../../types/config.js';
 import { type Input, toDeferResolve, toEntry } from '../../util/input.js';
-import { dirname, isInternal, join, toAbsolute } from '../../util/path.js';
-import { hasDependency, load } from '../../util/plugin.js';
+import { isInternal, join, toAbsolute } from '../../util/path.js';
+import { hasDependency } from '../../util/plugin.js';
+import { getReportersDependencies, resolveExtensibleConfig } from './helpers.js';
 import type { JestConfig, JestInitialOptions } from './types.js';
 
 // https://jestjs.io/docs/configuration
@@ -16,19 +17,6 @@ const isEnabled: IsPluginEnabled = ({ dependencies, manifest }) =>
 const config = ['jest.config.{js,ts,mjs,cjs,json}', 'package.json'];
 
 const entry = ['**/__tests__/**/*.[jt]s?(x)', '**/?(*.)+(spec|test).[jt]s?(x)'];
-
-const resolveExtensibleConfig = async (configFilePath: string) => {
-  let config = await load(configFilePath);
-  if (config?.preset) {
-    const { preset } = config;
-    if (isInternal(preset)) {
-      const presetConfigPath = toAbsolute(preset, dirname(configFilePath));
-      const presetConfig = await resolveExtensibleConfig(presetConfigPath);
-      config = Object.assign({}, presetConfig, config);
-    }
-  }
-  return config;
-};
 
 const resolveDependencies = async (config: JestInitialOptions, options: PluginOptions): Promise<Input[]> => {
   const { configFileDir } = options;
@@ -65,40 +53,7 @@ const resolveDependencies = async (config: JestInitialOptions, options: PluginOp
         ? [config.testEnvironment]
         : [];
   const resolvers = config.resolver ? [config.resolver] : [];
-
-  // Resolve dependencies for jest-junit reporter config
-  const jUnitReporterDeps: Array<string> = [];
-  for (const reporter of config.reporters ?? []) {
-    if (typeof reporter !== 'string' && reporter[0] === 'jest-junit') {
-      const {
-        testCasePropertiesFile,
-        testCasePropertiesDirectory,
-        testSuitePropertiesFile,
-        testSuitePropertiesDirectory,
-      } = reporter[1];
-      const testCaseFileName =
-        typeof testCasePropertiesFile === 'string' ? testCasePropertiesFile : 'junitProperties.js';
-      const testCaseDirectory =
-        typeof testCasePropertiesDirectory === 'string' ? testCasePropertiesDirectory : options.rootCwd;
-      const testSuiteFileName =
-        typeof testSuitePropertiesFile === 'string' ? testSuitePropertiesFile : 'junitTestCaseProperties.js';
-      const testSuiteDirectory =
-        typeof testSuitePropertiesDirectory === 'string' ? testSuitePropertiesDirectory : options.rootCwd;
-      const testCaseFilePath = join(testCaseDirectory, testCaseFileName);
-      const testSuiteFilePath = join(testSuiteDirectory, testSuiteFileName);
-      jUnitReporterDeps.push(testCaseFilePath);
-      jUnitReporterDeps.push(testSuiteFilePath);
-    }
-  }
-
-  let reporters = config.reporters
-    ? config.reporters
-        .map(reporter => (typeof reporter === 'string' ? reporter : reporter[0]))
-        .filter(reporter => !['default', 'github-actions', 'summary'].includes(reporter))
-    : [];
-
-  reporters = [...reporters, ...jUnitReporterDeps];
-
+  const reporters = getReportersDependencies(config, options);
   const watchPlugins =
     config.watchPlugins?.map(watchPlugin => (typeof watchPlugin === 'string' ? watchPlugin : watchPlugin[0])) ?? [];
   const transform = config.transform
